requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorclass==2.2.2
distlib==0.3.4
Django==4.0.4
docopt==0.6.2
filelock==3.7.0
idna==3.4
kombu==5.2.4
libvirt-python==8.3.0
packaging==22.0
pip-upgrader==1.4.15
pipe==2.0
platformdirs==2.5.2
prompt-toolkit==3.0.36
pytz==2022.7
redis==4.4.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
terminaltables==3.1.10
urllib3==1.26.13
vine==5.0.0
virtualenv==20.14.1
wcwidth==0.2.5
